+++ template/js/p/componentTemplate.js
@@ -1,17 +1,26 @@
 /**
+/**
  *
  * @param {string} Component component name
  * @param { boolean } cssmodule whether to use cssmodule or not
+ * @param { boolean } styled whether to use styled components or not
  */
-function componentTemplate(Component, cssmodule) {
+function componentTemplate(Component, cssmodule, styled) {
   return (
     `import React from 'react'
-${cssmodule ? `import cStyle from './${Component}.scss'`: `import './${Component}.scss'` }
+${styled ? `import styled from 'styled-components'` : cssmodule ? `import cStyle from './${Component}.scss'`: `import './${Component}.scss'`}
+${styled ? `
+const ${Component}Container = styled.div${`\`
 
+\``}
+` : ''}
 export class ${Component} extends React.PureComponent {
   render() {
     return (
-      <div className=${cssmodule ? `{cStyle.${Component}}` : `'${Component}'`}>${Component}</div>
+      ${  styled
+        ? `<${Component}Container>${Component}</${Component}Container>`
+        : `<div className=${cssmodule ? `{cStyle.${Component}}` : `'${Component}'`}>${Component}</div>`
+      }
     )
   }
 }