+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## 0.1.2
+
+# 0.2.0
+
+- Add styled components support
+- `-m` opition no long need to be true/false,if u want to use css module, just type `create-react-component -m MyComponent`
+
+# 0.1.2
 
 Add `-m [true|false]` option to specify whether to use `css module` or not.